requirements.txt
python-dotenv==1.0.1
pre-commit==3.5.0
fastapi==0.115.7
starlette
starlette-context==0.3.6
uvicorn==0.34.0
gunicorn==23.0.0
requests==2.32.3
pymongo==4.11
motor==3.7.0
pytest==8.3.4
pytest-cov==6.0.0
concurrent-log-handler==0.9.25
slowapi==0.1.9
asyncio==3.4.3
redis==6.2.0
PyJWT
python-jose
passlib[bcrypt]
black
isort
python-multipart
ruff
pre-commit
python-dateutil
pydantic[email]
pymsteams
httpx